updating schedule_data allowing package to install
Right now installation fails when prepping data for lazyloading because the API pull from myfantasyleague.com was not working correctly. This fixes that and an issue related to pulling the kickoff time from the superbowl (the kickoff time is both the first and the last game now). This fix should be good for the offseason / until the 2020 schedule function is ready to go. Tested this & the other pull request by downloading the package and installing it locally and it worked great.